feat: refactor color utility functions and update Button component styles
