+++ .storybook/preview.ts
@@ -1,5 +1,8 @@
 import type { Preview } from '@storybook/react';
 import '../lib/scss/main.scss';
+import { applyPalette } from '../lib/utils/colors.utils';
+
+applyPalette('#318CE7');
 
 const preview: Preview = {
   parameters: {
