+++ global.d.ts
@@ -1,0 +1,5 @@
+// declaration.d.ts
+declare module '*.scss' {
+  const content: Record<string, string>;
+  export default content;
+}
