+++ lib/components/button/button.module.scss
@@ -1,0 +1,21 @@
+.button {
+  background-color: var(--button-bg);
+  border: none;
+  border-radius: 4px;
+  color: var(--button-color);
+  padding-inline: 1rem;
+  height: var(--button-height, 2rem);
+  cursor: pointer;
+  &:hover {
+    background-color: var(--button-bg-hover);
+  }
+
+  &:active {
+    background-color: var(--button-bg);
+  }
+
+  &:focus {
+    outline: 1px solid var(--button-outline-focus-color);
+    outline-offset: 2px;
+  }
+}
