+++ lib/components/button/button.tsx
@@ -1,8 +1,14 @@
+import styles from './button.module.scss';
+
 export type ButtonProps = {
   onClick: () => void;
   label: string;
 };
 
 export function Button(props: ButtonProps) {
-  return <button onClick={props.onClick}>{props.label}</button>;
+  return (
+    <button className={styles.button} onClick={props.onClick}>
+      {props.label}
+    </button>
+  );
 }
