--- src/components/button/button.tsx
@@ -1,8 +1,0 @@
-export interface ButtonProps {
-  label: string;
-  onClick: () => void;
-}
-
-export const Button: React.FC<ButtonProps> = ({ label, onClick }) => {
-  return <button onClick={onClick}>{label}</button>;
-};
