--- src/index.scss
@@ -1,1 +1,0 @@
-@forward 'scss/main.scss';
