--- src/index.tsx
@@ -1,2 +1,0 @@
-export { Button } from './components/button/button'; // Exports the Button component
-export type { ButtonProps } from './components/button/button'; // Explicitly export types for Button props
